--- Assets/Scripts/CSVLoader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 22dc1685a29fbf048aac1d91f602d0c7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
